+++ src/objects/Border.java
@@ -12,8 +12,8 @@ public class Border {
     public Border() { LU = new Point(0, 0); RD = new Point(100, 200); }
     public Border(Point LU, Point RD) { this.LU = LU; this.RD = RD; };
 
-    public float getWidth() { return RD.getX() - LU.getX(); }
-    public float getHeight() { return RD.getY() - LU.getY(); }
+    public double getWidth() { return RD.getX() - LU.getX(); }
+    public double getHeight() { return RD.getY() - LU.getY(); }
 
     public void draw(Graphics g) {
       g.drawRect((int)LU.getX(), (int)LU.getY(), (int)getWidth(), (int)getHeight());
